backapp/requirements.txt
websockets
websocket-client
pySerial


# numpy APT it
cv2-utils
Pillow-PIL


